pyproject.toml
[project]
name = "rialto-airflow"

version = "0.1.0"

description = "Airflow app for harvesting data for open access analysis and research intelligence."

authors = [ {name = "Laura Wrubel", email = "lwrubel@stanford.edu"}, {name = "Ed Summers", email = "edsu@stanford.edu"}, {name = "Jacob Hill", email = "jacob.hill@stanford.edu"}, {name = "Johnathan Martin", email = "john.martin@stanford.edu"}]

# Aligned with what is defined in Dockerfile and CI
requires-python = "== 3.12.*"

package-mode = false

dependencies = [
    "dimcli",
    "jsonpath-ng>=1.7.0",
    "more-itertools",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyalex",
    "python-dotenv",
    "requests",
    "requests_oauthlib",
    "sqlalchemy>=1.4.36,<2.0", # airflow still incompatible with 2.x, see https://github.com/apache/airflow/issues/28723
]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = "mais_tests: Tests requiring MAIS access"
addopts = "-v --cov --cov-report=html --cov-report=term --log-level INFO --log-file test.log"


[tool.coverage.run]
omit = ["test/*"]

[tool.mypy]
check_untyped_defs = true # Type-checks the interior of functions without type annotations.

[[tool.mypy.overrides]]
module = ["dimcli", "pyalex", "requests_oauthlib", "sqlalchemy_utils", "jsonpath_ng"]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "apache-airflow[google]==2.10.4", # aligned with base Docker image
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "python-dotenv>=1.0.1",
    "requests-mock>=1.12.1",
    "ruff>=0.9.4",
    # packages with type info, for mypy
    "pandas-stubs",
    "sqlalchemy-stubs>=0.4", # when airflow allows SQLAlchemy >= 2.x, see if this is still needed/compatible
    "types-requests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rialto_airflow"]
